refactor(audio): clarify source bridge API
