--- examples/__init__.py
@@ -1,1 +1,0 @@
-"""Usage examples for aiosendspin."""
